requirements.txt
# AI 内容生成
openai>=1.0.0

# 数据采集
ccxt>=4.0.0         # 采集币圈行情
yfinance>=0.2.0     # 采集美股行情

# HTML 模板引擎
jinja2>=3.1.0       # 批量生成网页的核心

# 工具库
requests>=2.31.0
python-dotenv>=1.0.0
